requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
numpy
opencv-python
scenedetect
paddleocr
pillow
librosa
soundfile
pyloudnorm
faster-whisper
requests
openai>=1.42.0
python-dotenv
yt-dlp
redis
SQLAlchemy
psycopg2-binary
boto3
reportlab
sentry-sdk
prometheus-client
gspread
google-auth
